+++ History.md
@@ -1,7 +1,15 @@
+## 1.0.3 / 28-09-2016
+
+- fix(#95): better image type inference and support check.
+- fix(background): pass proper background RGB color for PNG image conversion.
+- feat(types): validate supported image types by current `libvips` compilation.
+- feat(types): consistent SVG image checking.
+- feat(api): add public functions `VipsIsTypeSupported()`, `IsImageTypeSupportedByVips()` and `IsSVGImage()`.
+
 ## 1.0.2 / 27-09-2016
 
-- feat(#95): support GIF, SVG and PDF formats
-- fix(#108): auto-width and height calculations now round instead of floor
+- feat(#95): support GIF, SVG and PDF formats.
+- fix(#108): auto-width and height calculations now round instead of floor.
 
 ## 1.0.1 / 22-06-2016